storage/pyproject.toml
[project]
name = "llmdocumentstore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "langchain-openai>=0.3.33",
    "mistralai>=1.9.10",
    "pydantic>=2.11.9",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "aioboto3>=15.1.0",
]
